+++ LocalPackages/RTSComponentKit/Sources/RTSComponentKit/Manager/SubscriptionManager.swift
@@ -7,28 +7,35 @@ import AVFoundation
 import Foundation
 import MillicastSDK
 
-public final class SubscriptionManager: ObservableObject {
+protocol SubscriptionManagerDelegate: AnyObject {
+    func onSubscribed()
+    func onSubscribedError(_ reason: String)
+    func onVideoTrack(_ track: MCVideoTrack, withMid mid: String)
+    func onAudioTrack(_ track: MCAudioTrack, withMid mid: String)
+    func onConnected()
+    func onStreamActive()
+    func onStreamInactive()
+    func onStreamStopped()
+    func onConnectionError(reason: String)
+}
 
-    private let listener: MCSubscriberListener
+final class SubscriptionManager {
 
     private let queueLabelKey = DispatchSpecificKey<String>()
     private let queueSub = DispatchQueue(label: "mc-QSub", qos: .userInitiated)
-
-    init(listener: MCSubscriberListener) {
-        self.listener = listener
-    }
-
     private var subscriber: MCSubscriber?
 
-    public func enableAudio(for track: MCAudioTrack?, enable: Bool) {
+    weak var delegate: SubscriptionManagerDelegate?
+
+    func enableAudio(for track: MCAudioTrack?, enable: Bool) {
         track?.enable(enable)
     }
 
-    public func enableVideo(for track: MCVideoTrack?, enable: Bool) {
+    func enableVideo(for track: MCVideoTrack?, enable: Bool) {
         track?.enable(enable)
     }
 
-    public func renderAudioTrack(_ track: MCAudioTrack?) async {
+    func renderAudioTrack(_ track: MCAudioTrack?) async {
         let task = {
             guard track != nil else {
                 return
@@ -39,7 +46,7 @@ public final class SubscriptionManager: ObservableObject {
         runOnQueue(log: "Render subscribe audio", task, queueSub)
     }
 
-    public func setAudioTrackVolume(_ volume: Double, audioTrack: MCAudioTrack) -> Bool {
+    func setAudioTrackVolume(_ volume: Double, audioTrack: MCAudioTrack) -> Bool {
         guard let subscriber = subscriber, subscriber.isSubscribed() else {
             return false
         }
@@ -47,7 +54,7 @@ public final class SubscriptionManager: ObservableObject {
         return true
     }
 
-    public func connect() async -> Bool {
+    func connect() async -> Bool {
         let task = { [weak self] in
             guard let self = self, let subscriber = self.subscriber else {
                 return
@@ -66,7 +73,7 @@ public final class SubscriptionManager: ObservableObject {
         return true
     }
 
-    public func connect(streamName: String, accountID: String) async -> Bool {
+    func connect(streamName: String, accountID: String) async -> Bool {
         let task = { [weak self] in
             guard let self = self, let subscriber = self.makeSubscriber() else {
                 return
@@ -91,7 +98,7 @@ public final class SubscriptionManager: ObservableObject {
         return true
     }
 
-    public func startSubscribe() async -> Bool {
+    func startSubscribe() async -> Bool {
         let task = { [weak self] in
             guard let self = self, let subscriber = self.subscriber else {
                 return
@@ -108,7 +115,7 @@ public final class SubscriptionManager: ObservableObject {
         return true
     }
 
-    public func stopSubscribe() async -> Bool {
+    func stopSubscribe() async -> Bool {
         let task = { [self] in
             guard let subscriber = subscriber, isSubscribed else {
                 return
@@ -131,7 +138,7 @@ public final class SubscriptionManager: ObservableObject {
         return true
     }
 
-    public func startRender(of track: MCVideoTrack?, on renderer: MCIosVideoRenderer) async {
+    func startRender(of track: MCVideoTrack?, on renderer: MCIosVideoRenderer) async {
         let task = {
             guard let track = track else {
                 return
@@ -141,7 +148,7 @@ public final class SubscriptionManager: ObservableObject {
         runOnQueue(log: "Render subscribe video", task, queueSub)
     }
 
-    public func stopRender(of track: MCVideoTrack?, on renderer: MCIosVideoRenderer) async {
+    func stopRender(of track: MCVideoTrack?, on renderer: MCIosVideoRenderer) async {
         let task = {
             guard let track = track else {
                 return
@@ -160,7 +167,7 @@ private extension SubscriptionManager {
         guard let subscriber = MCSubscriber.create() else {
             return nil
         }
-        subscriber.setListener(listener)
+        subscriber.setListener(self)
 
         return subscriber
     }
@@ -180,6 +187,65 @@ private extension SubscriptionManager {
         credentials.apiUrl = "https://director.millicast.com/api/director/subscribe"
 
         return credentials
+    }
+}
+
+extension SubscriptionManager: MCSubscriberListener {
+
+    func onSubscribed() {
+        delegate?.onSubscribed()
+    }
+
+    func onSubscribedError(_ reason: String!) {
+        delegate?.onSubscribedError(reason)
+    }
+
+    func onVideoTrack(_ track: MCVideoTrack!, withMid mid: String!) {
+        delegate?.onVideoTrack(track, withMid: mid)
+    }
+
+    func onAudioTrack(_ track: MCAudioTrack!, withMid mid: String!) {
+        delegate?.onAudioTrack(track, withMid: mid)
+    }
+
+    func onActive(_ streamId: String!, tracks: [String]!, sourceId: String!) {
+        delegate?.onStreamActive()
+    }
+
+    func onInactive(_ streamId: String!, sourceId: String!) {
+        delegate?.onStreamInactive()
+    }
+
+    func onStopped() {
+        delegate?.onStreamStopped()
+    }
+
+    func onVad(_ mid: String!, sourceId: String!) {
+        // TODO:
+    }
+
+    func onLayers(_ mid: String!, activeLayers: [MCLayerData]!, inactiveLayers: [MCLayerData]!) {
+        // TODO:
+    }
+
+    func onConnected() {
+        delegate?.onConnected()
+    }
+
+    func onConnectionError(_ status: Int32, withReason reason: String!) {
+        delegate?.onConnectionError(reason: reason)
+    }
+
+    func onSignalingError(_ message: String!) {
+        // TODO:
+    }
+
+    func onStatsReport(_ report: MCStatsReport!) {
+        // TODO:
+    }
+
+    func onViewerCount(_ count: Int32) {
+        // TODO:
     }
 }
