+++ LocalPackages/RTSComponentKit/Sources/RTSComponentKit/Store/RTSDataStore.swift
@@ -28,25 +28,19 @@ public final class RTSDataStore: ObservableObject {
     @Published public private(set) var isSubscribeVideoEnabled: Bool = true
 
     private let subscriptionManager: SubscriptionManager
-    private let listener: MCSubscriberListener
     private let subscriptionVideoRenderer: MCIosVideoRenderer
     private var audioTrack: MCAudioTrack?
     private var videoTrack: MCVideoTrack?
 
-    init(subscriptionManager: SubscriptionManager, listener: MCSubscriberListener, subscriptionVideoRenderer: MCIosVideoRenderer) {
+    init(subscriptionManager: SubscriptionManager, subscriptionVideoRenderer: MCIosVideoRenderer) {
         self.subscriptionManager = subscriptionManager
-        self.listener = listener
         self.subscriptionVideoRenderer = subscriptionVideoRenderer
+        self.subscriptionManager.delegate = self
     }
 
     public convenience init() {
-        let listener = SubscriptionListener()
-        self.init(
-            subscriptionManager: SubscriptionManager(listener: listener),
-            listener: listener,
-            subscriptionVideoRenderer: MCIosVideoRenderer(colorRangeExpansion: false)
-        )
-        listener.delegate = self
+        self.init(subscriptionManager: SubscriptionManager(), subscriptionVideoRenderer: MCIosVideoRenderer())
+        self.subscriptionManager.delegate = self
     }
 
     // MARK: Subscribe API methods
@@ -135,9 +129,10 @@ public final class RTSDataStore: ObservableObject {
     }
 }
 
-// MARK: SubscriptionListenerDelegate implementation
+// MARK: SubscriptionManagerDelegate implementation
 
-extension RTSDataStore: SubscriptionListenerDelegate {
+extension RTSDataStore: SubscriptionManagerDelegate {
+
     func onStreamActive() {
         Task {
             await MainActor.run {
